Test payment gateway, spring profile added

+++ src/main/java/com/coconet/model/Subscription.java
@@ -5,6 +5,7 @@ import org.hibernate.validator.constraints.NotEmpty;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.sql.Time;
 import java.util.Date;
 
@@ -29,23 +30,23 @@ public class Subscription {
     private Time preferredDeliveryTime;
     @NotNull
     @Column(name = "quantity_per_day")
-    private double quantityPerDay;
+    private BigDecimal quantityPerDay;
     @NotNull
     @Column(name = "total_quantity")
-    private double totalQuantity;
+    private BigDecimal totalQuantity;
     @Column(name = "discount_percentage")
     private int discountPercentage;
     @Column(name = "discount")
-    private double discount;
+    private BigDecimal discount;
     @NotNull
     @Column(name = "total_number_of_days")
     private int totalNumberOfDays;
     @NotNull
     @Column(name = "actual_price")
-    private double actualPrice;
+    private BigDecimal actualPrice;
     @NotNull
     @Column(name = "total_price")
-    private double totalPrice;
+    private BigDecimal totalPrice;
     @NotNull
     @Column(name = "payment_status")
     private String paymentStatus;
@@ -102,11 +103,11 @@ public class Subscription {
         this.preferredDeliveryTime = preferredDeliveryTime;
     }
 
-    public double getQuantityPerDay() {
+    public BigDecimal getQuantityPerDay() {
         return quantityPerDay;
     }
 
-    public void setQuantityPerDay(double quantityPerDay) {
+    public void setQuantityPerDay(BigDecimal quantityPerDay) {
         this.quantityPerDay = quantityPerDay;
     }
 
@@ -118,19 +119,19 @@ public class Subscription {
         this.totalNumberOfDays = totalNumberOfDays;
     }
 
-    public double getActualPrice() {
+    public BigDecimal getActualPrice() {
         return actualPrice;
     }
 
-    public void setActualPrice(double actualPrice) {
+    public void setActualPrice(BigDecimal actualPrice) {
         this.actualPrice = actualPrice;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
@@ -150,11 +151,11 @@ public class Subscription {
         this.subscriptionPlan = subscriptionPlan;
     }
 
-    public double getTotalQuantity() {
+    public BigDecimal getTotalQuantity() {
         return totalQuantity;
     }
 
-    public void setTotalQuantity(double totalQuantity) {
+    public void setTotalQuantity(BigDecimal totalQuantity) {
         this.totalQuantity = totalQuantity;
     }
 
@@ -166,11 +167,11 @@ public class Subscription {
         this.discountPercentage = discountPercentage;
     }
 
-    public double getDiscount() {
+    public BigDecimal getDiscount() {
         return discount;
     }
 
-    public void setDiscount(double discount) {
+    public void setDiscount(BigDecimal discount) {
         this.discount = discount;
     }
